More efficiently delete all settings for given key, if cleaning results in all properties needing to be deleted.

+++ net.solarnetwork.node.settings.ca/src/net/solarnetwork/node/settings/ca/CASettingsService.java
@@ -575,10 +575,20 @@ public class CASettingsService implements SettingsService, BackupResourceProvide
 						}
 					}
 				}
-				for ( String key : keysToRemove ) {
-					propUpdates.remove(key);
-					if ( !configurationOnly ) {
-						settingDao.deleteSetting(settingKey, key);
+				if ( !keysToRemove.isEmpty() ) {
+					if ( keysToRemove.size() == propUpdates.size() ) {
+						// delete all optimization
+						propUpdates.clear();
+						if ( !configurationOnly ) {
+							settingDao.deleteSetting(settingKey);
+						}
+					} else {
+						for ( String key : keysToRemove ) {
+							propUpdates.remove(key);
+							if ( !configurationOnly ) {
+								settingDao.deleteSetting(settingKey, key);
+							}
+						}
 					}
 				}
 			}
